requirements.txt
fastapi
uvicorn
requests
python-dotenv
streamlit
pydantic
groq-python
instructor
groq